requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
groq==1.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
jiter==0.13.0
langdetect==1.0.9
packaging==26.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-telegram-bot==20.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
